separation des test creation depuis un tableau
